renamed. battery level now updates, if possible
previously battery percentage would only update when .zshrc was sourced. Changed double quotes to single to force evaluation on each command.
Because not all computers are mac (huh) or have access to battery statistics, so modified so that if it can't check battery it will omit that from the prompt.
Forced the prompt assignment onto one line by using a newline character and forcing interpolation with $''
Colored the time stamp brackets so that they're more noticeable
Renamed!

Former-commit-id: 86fc00edfec56f1dcae375c9986818c9ee206e91